BZ1 and n64 support
BZ1 support added.  If a BZN is opened while the BZ1 mode check box in
the menu is checked, it will assume there is a padding byte between type
and size allowing BZ1 BZNs to load.

If a BIN is loaded, Battlzone64 logic is used.  These fields are not
typed and are big endian (for multi-byte data types).